use session.user for requests

+++ db/cart.js
@@ -2,9 +2,9 @@ const db = require('./index.js');
 
 //gets the total of each set of items and returns the total price of each set of items
 const getCartProducts = (req, res, next) => {
-    const { userId } = req.body;
-
-    if (!userId) return res.status(400).send('400 ERROR: User not specified');
+    if (!req.user) return res.status(401).send('User not logged in.');
+    
+    const userId = req.user.id;
 
     const query = `SELECT product_id, quantity, name, SUM(price) AS total FROM cart_products
     JOIN carts
@@ -26,9 +26,9 @@ const getCartProducts = (req, res, next) => {
 
 //gets the total number of items in the cart and the total price of the cart
 const getCartTotal = (req, res, next) => {
-    const { userId } = req.body;
-
-    if (!userId) return res.status(400).send('400 ERROR: User not specified');
+    if (!req.user) return res.status(401).send('User not logged in.');
+    
+    const userId = req.user.id;
 
     const query = `SELECT COUNT (product_id) AS product_types, SUM (product_id) AS total_items, SUM(price) AS total FROM cart_products
     JOIN carts
@@ -49,7 +49,10 @@ const getCartTotal = (req, res, next) => {
 
 //removes a product from the cart
 const removeProductFromCart = (req, res, next) => {
-    const { userId, productId } = req.body;
+    if (!req.user) return res.status(401).send('User not logged in.');
+    
+    const { productId } = req.body;
+    const userId = req.user.id;
 
     const query = `DELETE FROM cart_products
     USING carts, products
@@ -72,9 +75,9 @@ const removeProductFromCart = (req, res, next) => {
 
 // clears all items from the current user's cart
 const clearCartItems = (req, res, next) => {
-    const { userId } = req.body;
-
-    if (!userId) return res.status(400).send('400 ERROR: User not specified');
+    if (!req.user) return res.status(401).send('User not logged in.');
+    
+    const userId = req.user.id;
 
     const query = `DELETE FROM cart_products 
     USING carts
@@ -92,9 +95,9 @@ const clearCartItems = (req, res, next) => {
 
 // Checks to make sure all products in the cart are in sufficient stock to create an order.
 const checkCartProductsStock = (req, res, next) => {
-    const { userId } = req.body;
-
-    if (!userId) return res.status(400).send('400 ERROR: User not specified');
+    if (!req.user) return res.status(401).send('User not logged in.');
+    
+    const userId = req.user.id;
 
     const query = `SELECT product_id, quantity, stock
         FROM cart_products
@@ -117,9 +120,9 @@ const checkCartProductsStock = (req, res, next) => {
 
 // Creates an order entry attached to the current user and transfers the items from the associated entries on the cart_products table to the associated order on the order_products table
 const checkoutCart = (req, res, next) => {
-    const { userId } = req.body;
-
-    if (!userId) return res.status(400).send('400 ERROR: User not specified');
+    if (!req.user) return res.status(401).send('User not logged in.');
+    
+    const userId = req.user.id;
 
     const query = `WITH current_user_cart_products AS (
         DELETE FROM cart_products
